chore: renames files to correct casing
