+++ packages/datepickers/demo/date_PickerRange.stories.mdx
@@ -1,7 +1,7 @@
 import { Meta, ArgsTable, Canvas, Story, Markdown } from '@storybook/addon-docs';
 import { useArgs } from '@storybook/client-api';
 import { DatePickerRange } from '@zendeskgarden/react-datepickers';
-import { DatePickerRangeStory } from './stories/Date_PickerRangeStory';
+import { DatePickerRangeStory } from './stories/DatePickerRangeStory';
 import { DATE_STYLE_OPTIONS } from './stories/data';
 import README from '../README.md';
