+++ packages/datepickers/src/elements/DatePickerRange/DatePickerRange.spec.tsx
@@ -13,7 +13,7 @@ import { subDays } from 'date-fns/subDays';
 import { addMonths } from 'date-fns/addMonths';
 import { subMonths } from 'date-fns/subMonths';
 import mockDate from 'mockdate';
-import { DatePickerRange } from './Date_PickerRange';
+import { DatePickerRange } from './DatePickerRange';
 import { IDatePickerRangeProps } from '../../types';
 
 const DEFAULT_START_VALUE = new Date(2019, 1, 5);
